Pad out colgroup with columns after last entry in columnStyles
so that the `<colgroup>` definition describes all columns of the table, and passes HTML validation.
