+++ lib/reports/format/MovePdfGenerator.js
@@ -7,7 +7,6 @@ import PdfPrinter from 'pdfmake';
 
 import { ORG_NAME, PageOrientation } from '@/lib/Constants';
 import NumberFormatters from '@/lib/i18n/NumberFormatters';
-import ArrayStats from '@/lib/math/ArrayStats';
 import FormatCss from '@/lib/reports/format/FormatCss';
 import TableUtils from '@/lib/reports/format/TableUtils';
 import BarChartGenerator from '@/lib/reports/svg/BarChartGenerator';
@@ -259,37 +258,6 @@ class MovePdfGenerator {
 
   // ReportBlock.TABLE
 
-  static getNumSectionColumns(section) {
-    if (section.length === 0) {
-      /*
-       * This can happen for sections that aren't provided in the layout.  For
-       * instance, the header and footer are both optional, and default to the
-       * empty list if not provided.
-       */
-      return 0;
-    }
-    /*
-     * Since we define the table layouts ourselves, we can take the shortcut
-     * of assuming that these layouts specify an equal number of columns in each
-     * row.  Given that, we only need to check the first row.
-     */
-    const [firstRow] = section;
-    return ArrayStats.sum(
-      firstRow.map(({ colspan = 1 }) => colspan),
-    );
-  }
-
-  static getNumTableColumns(header, body, footer) {
-    const numHeaderColumns = MovePdfGenerator.getNumSectionColumns(header);
-    const numBodyColumns = MovePdfGenerator.getNumSectionColumns(body);
-    const numFooterColumns = MovePdfGenerator.getNumSectionColumns(footer);
-    return Math.max(
-      numHeaderColumns,
-      numBodyColumns,
-      numFooterColumns,
-    );
-  }
-
   static getColumnWidths(columnStyles, numColumns) {
     const widths = new Array(numColumns).fill('auto');
     columnStyles.forEach(({ c, style = {} }) => {
@@ -458,7 +426,7 @@ class MovePdfGenerator {
         text: caption,
       });
     }
-    const numColumns = MovePdfGenerator.getNumTableColumns(header, body, footer);
+    const numColumns = TableUtils.getNumTableColumns(header, body, footer);
     const tableRows = MovePdfGenerator.getTableRows(
       header,
       body,
